Refactored to a happy place.
Remaining work is just to complete the test states and events...
