+++ cypress/.eslintrc.json
@@ -1,0 +1,3 @@
+{
+  "extends": ["plugin:cypress/recommended"]
+}
