+++ package.json
@@ -6,8 +6,8 @@
   "scripts": {
     "build": "NODE_ENV=production parcel build index.html",
     "start": "parcel index.html",
-    "cypress": "cypress run",
-    "cypress:open": "cypress open",
+    "cypress": "cypress open",
+    "cypress:run": "cypress run --headless",
     "cypress:chrome": "cypress open --browser chrome"
   },
   "repository": {
